#78 Initial Refactoring
Got rid of as many static variables as i could cause I was not using them properly. The code was so intertwined because of the static variables that I ended up having to change an absurd amount of files. Obviously there is much more to do, but at this point it compiles, unsure if it is functional at all